todo/requirements.txt
Django==1.6.1
Pygments==1.6
South==0.8.4
djangorestframework==2.3.10
